# openapi.json
{
  "openapi" : "3.1.0",
  "components" : {
    "schemas" : {
      "AddNamespaceRequest" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string"
          }
        }
      },
      "Cluster" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "server" : {
            "type" : "string"
          },
          "user" : {
            "type" : "string"
          },
          "connected" : {
            "type" : "boolean"
          }
        }
      }
    }
  },
  "paths" : {
    "/clusters" : {
      "get" : {
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Cluster"
                  }
                }
              }
            }
          }
        },
        "summary" : "List",
        "tags" : [ "Cluster Resource" ]
      }
    },
    "/clusters/{ctx}/namespaces" : {
      "get" : {
        "parameters" : [ {
          "name" : "ctx",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "type" : "string"
                  }
                }
              }
            }
          }
        },
        "summary" : "List",
        "tags" : [ "Namespace Resource" ]
      },
      "post" : {
        "parameters" : [ {
          "name" : "ctx",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AddNamespaceRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : { }
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "summary" : "Add",
        "tags" : [ "Namespace Resource" ]
      }
    },
    "/clusters/{ctx}/namespaces/discover" : {
      "post" : {
        "parameters" : [ {
          "name" : "ctx",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : { }
              }
            }
          }
        },
        "summary" : "Discover",
        "tags" : [ "Namespace Resource" ]
      }
    },
    "/clusters/{ctx}/namespaces/{ns}" : {
      "delete" : {
        "parameters" : [ {
          "name" : "ctx",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "ns",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : { }
              }
            }
          }
        },
        "summary" : "Remove",
        "tags" : [ "Namespace Resource" ]
      }
    },
    "/clusters/{ctx}/namespaces/{ns}/resources/{type}" : {
      "get" : {
        "parameters" : [ {
          "name" : "ctx",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "ns",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "type",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : { }
              }
            }
          }
        },
        "summary" : "List",
        "tags" : [ "Resource Resource" ]
      }
    },
    "/clusters/{ctx}/namespaces/{ns}/resources/{type}/{name}" : {
      "get" : {
        "parameters" : [ {
          "name" : "ctx",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "name",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "ns",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "type",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : { }
              }
            }
          }
        },
        "summary" : "Get",
        "tags" : [ "Resource Resource" ]
      }
    }
  },
  "info" : {
    "title" : "knals",
    "version" : "0.1.0-SNAPSHOT"
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Auto generated value"
  }, {
    "url" : "http://0.0.0.0:8080",
    "description" : "Auto generated value"
  } ]
}